exclude GUI folder from flake check

+++ RL_environment/dmcs_env.py
@@ -97,7 +97,7 @@ class DMControlEnvironment:
         return observation
 
     def step(
-        self, action: float
+        self, action: int
     ) -> tuple:  # todo check if action should be float or np.ndarray
         """Take a step in the environment.
